MUI4 fix. The list is not displayed when the window opens.
MUI_NListtree_Active is documented as:
Setting this attribute will move the cursor to the defined tree node
it is visible. If the node is in an opened tree the listview is
scrolling into the visible area.

So, I set it when the whole is visible. :-)

Hope it won't make it worse for MUI3...
Feel free to revert if you believe this change is bad.