Imap folders could appear twice.
When imap folders were created from the imap account configuration, the presence of imap folders  on the 2nd level or beyond could lead to double folders totally disturbing the folder tree if an order file existed (from which the folders are created in the first step, i.e., before the imap config is checked). SimpleMail currently supports only one level of imap folders and maps the subfolders to the 1st level. The workaround that is implemented attempts to load the true imap path from the .config file before the folder is added for an existing directory, in which case the folder normally is not added again.