Print available loggable modules when user query available ones.
The user can now query which modules have some logging information by
using ? as argument to the DEBUGMODULES argument. The list of modules
is built on compile time by placing a keyword extented string in a
dedicated section (that get removed during linking), which is then
extracted for each object file.

This approach may be extened later by function names or line numbers
as well.